+++ packages/orq-rc/RELEASES.md
@@ -4939,3 +4939,13 @@ Based on:
 - [typescript v4.0.0-rc.25] packages/orq-rc
 ### Releases
 - [NPM v4.0.0-rc.25] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.25 - packages/orq-rc
+
+## 2025-11-12 13:43:30
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.654.0 (2.750.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.0.0-rc.26] packages/orq-rc
+### Releases
+- [NPM v4.0.0-rc.26] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.26 - packages/orq-rc